fix: move configuration into one class
