+++ src/main/java/com/epam/spring/config/WebConfig.java
@@ -2,24 +2,19 @@ package com.epam.spring.config;
 
 import jakarta.servlet.ServletContext;
 import jakarta.servlet.ServletRegistration;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.web.WebApplicationInitializer;
 import org.springframework.web.context.ContextLoaderListener;
 import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
 import org.springframework.web.servlet.DispatcherServlet;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 
-@Configuration
 @EnableWebMvc
-@ComponentScan("com.epam.spring")
 public class WebConfig implements WebApplicationInitializer {
 
     @Override
     public void onStartup(ServletContext container) {
         AnnotationConfigWebApplicationContext context = new AnnotationConfigWebApplicationContext();
-        context.setConfigLocation("com.epam.spring");
-//        container.setInitParameter("spring.profiles.active", "dev");
+        context.setConfigLocation("com.epam.spring.config");
         container.addListener(new ContextLoaderListener(context));
         ServletRegistration.Dynamic dispatcher = container.addServlet("dispatcher", new DispatcherServlet(context));
         dispatcher.setLoadOnStartup(1);
